+++ libs/feature-rooms/src/lib/message-list/virtual-message-list/virtual-message-list.component.spec.ts
@@ -31,6 +31,7 @@ function msg(
     media: null,
     caption: null,
     captionHtml: null,
+    readReceipts: [],
   };
 }